+++ Source/Classes/AnimatedImages/PINAnimatedImageView.m
@@ -17,6 +17,7 @@
     CFTimeInterval _playHead;
     NSUInteger _playedLoops;
     NSUInteger _lastSuccessfulFrameIndex;
+    CFTimeInterval *_durations;
 }
 
 @property (nonatomic, assign) CGImageRef frameImage;
@@ -59,12 +60,45 @@
 {
     _animatedImage = animatedImage;
     _animatedImageRunLoopMode = NSRunLoopCommonModes;
+    _durations = NULL;
+    
+    if (animatedImage) {
+        [self initializeAnimatedImage:animatedImage];
+    }
+}
+
+- (void)initializeAnimatedImage:(nonnull PINCachedAnimatedImage *)animatedImage
+{
+    PINWeakify(self);
+    animatedImage.coverImageReadyCallback = ^(PINImage *coverImage) {
+        dispatch_async(dispatch_get_main_queue(), ^{
+            PINStrongify(self);
+            // In this case the lock is already gone we have to call the unlocked version therefore
+            [self coverImageCompleted:coverImage];
+        });
+    };
+    
+    animatedImage.playbackReadyCallback = ^{
+        dispatch_async(dispatch_get_main_queue(), ^{
+            // In this case the lock is already gone we have to call the unlocked version therefore
+            PINStrongify(self);
+            [self checkIfShouldAnimate];
+        });
+    };
+    if (animatedImage.playbackReady) {
+        [self checkIfShouldAnimate];
+    }
+  
+    [self resetDurationsWithAnimatedImage:animatedImage];
 }
 
 - (void)dealloc
 {
     if (_frameImage) {
         CGImageRelease(_frameImage);
+    }
+    if (_durations) {
+        free(_durations);
     }
 }
 
@@ -82,25 +116,7 @@
     _animatedImage = animatedImage;
     
     if (animatedImage != nil) {
-        PINWeakify(self);
-        animatedImage.coverImageReadyCallback = ^(PINImage *coverImage) {
-            dispatch_async(dispatch_get_main_queue(), ^{
-                PINStrongify(self);
-                // In this case the lock is already gone we have to call the unlocked version therefore
-                [self coverImageCompleted:coverImage];
-            });
-        };
-        
-        animatedImage.playbackReadyCallback = ^{
-            dispatch_async(dispatch_get_main_queue(), ^{
-                // In this case the lock is already gone we have to call the unlocked version therefore
-                PINStrongify(self);
-                [self checkIfShouldAnimate];
-            });
-        };
-        if (animatedImage.playbackReady) {
-            [self checkIfShouldAnimate];
-        }
+        [self initializeAnimatedImage:animatedImage];
     } else {
         // Clean up after ourselves.
         self.layer.contents = nil;
@@ -376,19 +392,37 @@
     }
 }
 
+- (void)resetDurationsWithAnimatedImage:(PINCachedAnimatedImage *)animatedImage
+{
+    PINAssertMain();
+    if (!animatedImage) {
+        return;
+    }
+    if (_durations) {
+        free(_durations);
+    }
+    _durations = malloc(sizeof(CFTimeInterval) * animatedImage.frameCount);
+    CFTimeInterval sum = 0.0f;
+    for (int i = 0; i < animatedImage.frameCount; i++) {
+        sum += [animatedImage durationAtIndex:i];
+        _durations[i] = sum;
+    }
+}
+
 - (NSUInteger)frameIndexAtPlayHeadPosition:(CFTimeInterval)playHead
 {
     PINAssertMain();
-    NSUInteger frameIndex = 0;
-    for (NSUInteger durationIndex = 0; durationIndex < self.animatedImage.frameCount; durationIndex++) {
-        playHead -= [self.animatedImage durationAtIndex:durationIndex];
-        if (playHead < 0) {
-            return frameIndex;
-        }
-        frameIndex++;
-    }
+    int low = 0, high = (int)_animatedImage.frameCount - 1;
     
-    return frameIndex;
+    while (low <= high) {
+        int mid = low + (high - low) / 2;
+        if (_durations[mid] < playHead) {
+            low = mid + 1;
+        } else {
+            high = mid - 1;
+        }
+    }
+    return MAX(MIN(low, (int)_animatedImage.frameCount - 1), 0);
 }
 
 @end