+++ Source/Classes/AnimatedImages/PINCachedAnimatedImage.h
@@ -58,7 +58,7 @@
  */
 @property (nonatomic, readonly) BOOL playbackReady;
 /**
- @abstract Return any error that has occured. Playback will be paused if this returns non-nil.
+ @abstract Return any error that has occurred. Playback will be paused if this returns non-nil.
  */
 @property (nonatomic, readonly) NSError *error;
 /**